+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 ### v0.6.x
-- add: Tutorial for saving live timeseries data (e.g., forecast) in context broker and timeseries database ([#336](https://github.com/RWTH-EBC/FiLiP/pull/363))
+- add: Unittest workflow via Github actions ([#388](https://github.com/RWTH-EBC/FiLiP/pull/388))
+
+### v0.6.2
+- add: New features to validate and update relationships for V2 entities ([#380](https://github.com/RWTH-EBC/FiLiP/pull/380))
+- add: ngsi-ld support for entity creation and update with keyvalues format ([#394](https://github.com/RWTH-EBC/FiLiP/pull/394))
+- fix: exception handling for special characters in attribute value  ([#381](https://github.com/RWTH-EBC/FiLiP/pull/381))
+- update: relax pydantic requirements ([#390](https://github.com/RWTH-EBC/FiLiP/pull/390))
+- update: pandas requirement from ~=1.3.5 to >=1.3.5,<2.3.0 ([#396](https://github.com/RWTH-EBC/FiLiP/pull/396))
+
+### v0.6.1
+- add: Tutorial for saving live timeseries data (e.g., forecast) in context broker and timeseries database ([#363](https://github.com/RWTH-EBC/FiLiP/pull/363))
 - add: Tutorial for using session object ([#370](https://github.com/RWTH-EBC/FiLiP/pull/370))
+- add: Validation of quotation mark in ``q`` strings for NGSI-LD ``get_entities_list`` ([#383](https://github.com/RWTH-EBC/FiLiP/pull/383/files))
 - fix: Addition of trailing slash if missing from base url ([#371](https://github.com/RWTH-EBC/FiLiP/pull/371))
 - fix: Serialization error for custom datamodels during batch operation ([#376](https://github.com/RWTH-EBC/FiLiP/pull/376))
 - update: omit invalid entities in `get_entity_list` ([#375](https://github.com/RWTH-EBC/FiLiP/pull/375))